restructure product route
