+++ src/api/controllers/index.js
@@ -1,4 +1,4 @@
 import userController from './user.controller.js'
-import productController from './product.controller.js'
+import productController from './products'
 
 export { userController, productController }
