+++ src/api/routes/v1/index.js
@@ -1,6 +1,6 @@
 import { Router } from 'express'
 import userRoute from './user.route.js'
-import productRoute from './product.route.js'
+import productRoute from './products'
 import apidocsRoute from './apidocs.route.js'
 
 const router = Router()
