+++ src/api/services/index.js
@@ -1,6 +1,6 @@
 import authService from './auth.service.js'
 import userService from './user.service.js'
 import tokenService from './token.service.js'
-import productService from './product.service.js'
+import productService from './products'
 
 export { authService, userService, tokenService, productService }
